feat: auto-balance brackets and quotes at end of completion text
When a completion text that ends with a opening bracket or quote
character is accepted, insert the corresponding closing character after
point.

This is particularly useful in Go where copilot often starts with just
suggesting the first line of a if statement, for loop, or function
defiition, along with the opening { character at the end of the line.

Accepting that is, yields to a unbalanced set of curly brackets in the
file, and other Emacs packages like smartparans tends to not be too
happy about it.

The auto-balancing feature is enabled/disabled with the
copilot-auto-balance custom variable. The auto-balancing pairs are
defined in the buffer-local variable copilot-auto-balance-pairs, which
defaults to:

    '(("(" . ")")
      ("[" . "]")
      ("{" . "}")
      ("\"" . "\"")
      ("'" . "'"))

As it's a buffer-local variable, it can easily be over-ridden in
major-mode hooks for different sets in different types of files. The
open and closing strings can also be longer than a single character if
need be.

By default, if the transform-fn given to copilot-accept-completion
returns a completion text that is shorter than the original (indicating
a partial accept), auto-balancing will not be performed.

This can be overridden though by setting copilot--auto-balance-partial
to t with a let block, from within which you call
copilot-accept-completion.

This is what copilot-accept-completion-by-line does, as it typically
makes sense to auto-balance when completing to the end of the line. As
opposed to copilot-accept-completion-by-word, which will only
auto-balance if the word being accepted is the very last piece of the
current completion.